Clean up DSP leaves and guard the biquad against bad bandwidth (#25)
- biquad_bpf: a non-positive bandwidth or sample rate made Q infinite or
  negative and collapsed the coefficients to a silent all-stop filter. Fall
  back to a unity passthrough and clear the delay line instead, so a
  misconfigured band leaves the signal observable rather than dead. Added a
  biquad_selftest case that pins the passthrough (goes red on the old path).
- modem_fsk: surface the IQ low-pass cutoff as modem_params_t.fsk_iq_lpf_hz
  (0 = built-in default / $FSK_IQ_LPF_HZ) instead of a getenv buried in a
  DSP leaf; the env override stays for offline sweeps.
- sw_nco: document that the per-sample cos/sin is deliberate — a phasor
  recurrence would break the chunk-invariant, byte-exact output the live
  decode and selftest rely on.
- modem_iq: drop a stale comment block describing a superseded per-sample
  differential the code no longer does, and fix the duplicated step numbers.
- iq_burst: reword the 1e-6 power floor comment so it doesn't oversell what
  is just a log10(0) guard.
- b210_rx_tx_core: retitle the header as the device-agnostic core it now is,
  and note that the IQ-level / burst snapshot getters are intentionally
  unsynchronised display-only reads.
- Make the (size_t)sps * 32 guards explicitly 64-bit (they already were —
  the cast binds before the multiply — but the width is now unambiguous).

The ASM_BIG_ENDIAN_U32 constant is already shared via src/dsp/asm_search.h.

Closes #25

+++ src/dsp/modem_iq.c
@@ -40,7 +40,7 @@ int modem_iq_to_bits(const int16_t *iq_pairs, size_t n_pairs,
         return -1;
     }
     int sps = p->samp_rate / p->bit_rate;
-    if (sps <= 1 || n_pairs < (size_t) sps * 32u) return -1;
+    if (sps <= 1 || n_pairs < (size_t) sps * (size_t) 32) return -1;
 
     // 1. Complex DC-block. 1-pole IIR HPF (α=0.995, ~76 Hz at 96 kHz)
     //    on I and Q separately. Removes the SDR's static DC bias which
@@ -121,15 +121,7 @@ int modem_iq_to_bits(const int16_t *iq_pairs, size_t n_pairs,
 #undef IQ_I
 #undef IQ_Q
 
-    // 3. Differential phase: y[k] = arg(z[k] * conj(z[k-1])). The first
-    //    output is at index 1; index 0 is undefined (no prev). Use atan2
-    //    so the result lies in [-π, π].
-    //
-    //    For ideal continuous-phase FSK at modulation index h, the
-    //    per-sample phase advance is ±πh/sps and integrates to ±πh per
-    //    symbol. At h=0.5 (MSK) a perfectly-strobed symbol differential
-    //    is ±π/2 — well separated from 0.
-    // 3. Symbol-rate differential phase at every sample position:
+    // 4. Symbol-rate differential phase at every sample position:
     //    dphi[i] = arg(z_mf[i+sps] · conj(z_mf[i]))
     //
     //    For MSK h=0.5 this is ±π/2 inside a symbol, with a smooth
